atfx_api/requirements.txt
allure-python-commons==2.9.45
atomicwrites==1.4.1
attrs==23.1.0
certifi==2022.12.7
chardet==3.0.4
colorama==0.4.6
cssselect2==0.7.0
distlib==0.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
Faker==18.4.0
idna==2.8
iniconfig==2.0.0
ipython-genutils==0.2.0
jdcal==1.4.1
jsonpath==0.82
loguru==0.7.0
lxml==4.9.2
numpy==1.24.3
openpyxl==3.0.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pluggy==0.13.1
py==1.11.0
pycryptodome==3.17
pymongo==3.11.4
PyMySQL==1.0.2
pytest==6.2.4
pytest-assume==2.4.3
pytest-check==2.1.4
python-dateutil==2.8.2
pytz==2023.3
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.0.2
reportlab==3.6.12
requests==2.22.0
shellescape==3.8.1
simplegeneric==0.8.1
six==1.16.0
svglib==1.5.1
tinycss2==1.2.1
toml==0.10.2
tzdata==2023.3
urllib3==1.25.11
wcwidth==0.1.7
webencodings==0.5.1
win32-setctime==1.1.0
WMI==1.4.9
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
zmq==0.0.0
